# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Save {
  id: number;
  userId: string;
  note?: string | null;
  url?: string | null;
  scrapedTitle?: string | null;
  description?: string | null;
  placeName?: string | null;
  countryCode?: string | null;
  lat?: number | null;
  lng?: number | null;
  tags?: string[] | null;
  category?: string | null;
  officialLink?: string | null;
  createdAt: string;
}

export interface UpdateSaveBody {
  scrapedTitle?: string | null;
  description?: string | null;
  placeName?: string | null;
  note?: string | null;
  tags?: string[] | null;
  category?: string | null;
  officialLink?: string | null;
}

export interface CreateSaveBody {
  note?: string | null;
  url?: string | null;
  scrapedTitle?: string | null;
  description?: string | null;
}

export type VerdictJsonType =
  (typeof VerdictJsonType)[keyof typeof VerdictJsonType];

export const VerdictJsonType = {
  choose: "choose",
  structure: "structure",
} as const;

export interface VerdictJson {
  type: VerdictJsonType;
  verdict: string;
  /**
   * @minItems 3
   * @maxItems 3
   */
  travelPatterns: string[];
  coreConflict: string;
  whatYoureMissing: string;
  whyThisFits: string;
  tradeoffs: string;
  avoidIf: string[];
  nextMove: string;
  /**
   * @minItems 3
   * @maxItems 3
   */
  anchors: string[];
  timingConfidence: string;
  stopDoingThis: string;
  usedSaveIds: number[];
}

export interface Decision {
  id: number;
  userId: string;
  question: string;
  result: string;
  resultJson?: VerdictJson | null;
  savesSnapshot: string;
  createdAt: string;
}

export interface CreateDecisionBody {
  question: string;
  saveIds: number[];
}

export interface SaveSelectionInput {
  question: string;
}

export interface SaveSelection {
  saveIds: number[];
}

export interface DeleteResult {
  success: boolean;
}

export interface ScrapeUrlBody {
  url: string;
}

export interface TagSaveBody {
  extractedText?: string | null;
}

export interface ScrapeResult {
  url: string;
  title?: string | null;
  image?: string | null;
  siteName?: string | null;
  extractedText?: string | null;
}