# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /decisions:
    get:
      operationId: listDecisions
      tags: [decisions]
      summary: List all decisions for the current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
    post:
      operationId: createDecision
      tags: [decisions]
      summary: Generate a travel decision from saves and a question
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDecisionBody"
      responses:
        "201":
          description: Decision created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /decisions/select-saves:
    post:
      operationId: selectSavesForDecision
      tags: [decisions]
      summary: AI-selects the most relevant saves for a given question
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSelectionInput"
      responses:
        "200":
          description: Selected save IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveSelection"
  /decisions/{id}:
    get:
      operationId: getDecision
      tags: [decisions]
      summary: Get a single decision
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
    delete:
      operationId: deleteDecision
      tags: [decisions]
      summary: Delete a decision
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    VerdictJson:
      type: object
      properties:
        type:
          type: string
          enum: [choose, structure]
        verdict:
          type: string
        travelPatterns:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 3
        coreConflict:
          type: string
        whatYoureMissing:
          type: string
        whyThisFits:
          type: string
        tradeoffs:
          type: string
        avoidIf:
          type: array
          items:
            type: string
        nextMove:
          type: string
        anchors:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 3
        timingConfidence:
          type: string
        stopDoingThis:
          type: string
        usedSaveIds:
          type: array
          items:
            type: integer
      required:
        - type
        - verdict
        - travelPatterns
        - coreConflict
        - whatYoureMissing
        - whyThisFits
        - tradeoffs
        - avoidIf
        - nextMove
        - anchors
        - timingConfidence
        - stopDoingThis
        - usedSaveIds
    Decision:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        question:
          type: string
        result:
          type: string
        resultJson:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/VerdictJson"
        savesSnapshot:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - question
        - result
        - savesSnapshot
        - createdAt
    CreateDecisionBody:
      type: object
      properties:
        question:
          type: string
        saveIds:
          type: array
          items:
            type: integer
      required:
        - question
        - saveIds
    SaveSelectionInput:
      type: object
      properties:
        question:
          type: string
      required:
        - question
    SaveSelection:
      type: object
      properties:
        saveIds:
          type: array
          items:
            type: integer
      required:
        - saveIds
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...