# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all saves for the current user
 */
export const ListSavesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  note: zod.string().nullish(),
  url: zod.string().nullish(),
  scrapedTitle: zod.string().nullish(),
  description: zod.string().nullish(),
  placeName: zod.string().nullish(),
  countryCode: zod.string().nullish(),
  lat: zod.number().nullish(),
  lng: zod.number().nullish(),
  tags: zod.array(zod.string()).nullish(),
  category: zod.string().nullish(),
  officialLink: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListSavesResponse = zod.array(ListSavesResponseItem);

/**
 * @summary Add a new travel save
 */
export const CreateSaveBody = zod.object({
  note: zod.string().nullish(),
  url: zod.string().nullish(),
  scrapedTitle: zod.string().nullish(),
  description: zod.string().nullish(),
});

/**
 * @summary Generate AI tags for a save
 */
export const TagSaveParams = zod.object({
  id: zod.coerce.number(),
});

export const TagSaveBody = zod.object({
  extractedText: zod.string().nullish(),
});

export const TagSaveResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  note: zod.string().nullish(),
  url: zod.string().nullish(),
  scrapedTitle: zod.string().nullish(),
  description: zod.string().nullish(),
  placeName: zod.string().nullish(),
  countryCode: zod.string().nullish(),
  lat: zod.number().nullish(),
  lng: zod.number().nullish(),
  tags: zod.array(zod.string()).nullish(),
  category: zod.string().nullish(),
  officialLink: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Geocode a save using AI + Nominatim
 */
export const GeocodeSaveParams = zod.object({
  id: zod.coerce.number(),
});

export const GeocodeSaveResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  note: zod.string().nullish(),
  url: zod.string().nullish(),
  scrapedTitle: zod.string().nullish(),
  description: zod.string().nullish(),
  placeName: zod.string().nullish(),
  countryCode: zod.string().nullish(),
  lat: zod.number().nullish(),
  lng: zod.number().nullish(),
  tags: zod.array(zod.string()).nullish(),
  category: zod.string().nullish(),
  officialLink: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update editable fields of a save
 */
export const UpdateSaveParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSaveBody = zod.object({
  scrapedTitle: zod.string().nullish(),
  description: zod.string().nullish(),
  placeName: zod.string().nullish(),
  note: zod.string().nullish(),
  tags: zod.array(zod.string()).nullish(),
  category: zod.string().nullish(),
  officialLink: zod.string().nullish(),
});

export const UpdateSaveResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  note: zod.string().nullish(),
  url: zod.string().nullish(),
  scrapedTitle: zod.string().nullish(),
  description: zod.string().nullish(),
  placeName: zod.string().nullish(),
  countryCode: zod.string().nullish(),
  lat: zod.number().nullish(),
  lng: zod.number().nullish(),
  tags: zod.array(zod.string()).nullish(),
  category: zod.string().nullish(),
  officialLink: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a save
 */
export const DeleteSaveParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteSaveResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all decisions for the current user
 */
export const listDecisionsResponseResultJsonOneTravelPatternsMin = 3;
export const listDecisionsResponseResultJsonOneTravelPatternsMax = 3;

export const listDecisionsResponseResultJsonOneAnchorsMin = 3;
export const listDecisionsResponseResultJsonOneAnchorsMax = 3;

export const ListDecisionsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  question: zod.string(),
  result: zod.string(),
  resultJson: zod
    .object({
      type: zod.enum(["choose", "structure"]),
      verdict: zod.string(),
      travelPatterns: zod
        .array(zod.string())
        .min(listDecisionsResponseResultJsonOneTravelPatternsMin)
        .max(listDecisionsResponseResultJsonOneTravelPatternsMax),
      coreConflict: zod.string(),
      whatYoureMissing: zod.string(),
      whyThisFits: zod.string(),
      tradeoffs: zod.string(),
      avoidIf: zod.array(zod.string()),
      nextMove: zod.string(),
      anchors: zod
        .array(zod.string())
        .min(listDecisionsResponseResultJsonOneAnchorsMin)
        .max(listDecisionsResponseResultJsonOneAnchorsMax),
      timingConfidence: zod.string(),
      stopDoingThis: zod.string(),
      usedSaveIds: zod.array(zod.number()),
    })
    .nullish(),
  savesSnapshot: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListDecisionsResponse = zod.array(ListDecisionsResponseItem);

/**
 * @summary Generate a travel decision from saves and a question
 */
export const CreateDecisionBody = zod.object({
  question: zod.string(),
  saveIds: zod.array(zod.number()),
});

/**
 * @summary AI-selects the most relevant saves for a given question
 */
export const SelectSavesForDecisionBody = zod.object({
  question: zod.string(),
});

export const SelectSavesForDecisionResponse = zod.object({
  saveIds: zod.array(zod.number()),
});

/**
 * @summary Get a single decision
 */
export const GetDecisionParams = zod.object({
  id: zod.coerce.number(),
});

export const getDecisionResponseResultJsonOneTravelPatternsMin = 3;
export const getDecisionResponseResultJsonOneTravelPatternsMax = 3;

export const getDecisionResponseResultJsonOneAnchorsMin = 3;
export const getDecisionResponseResultJsonOneAnchorsMax = 3;

export const GetDecisionResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  question: zod.string(),
  result: zod.string(),
  resultJson: zod
    .object({
      type: zod.enum(["choose", "structure"]),
      verdict: zod.string(),
      travelPatterns: zod
        .array(zod.string())
        .min(getDecisionResponseResultJsonOneTravelPatternsMin)
        .max(getDecisionResponseResultJsonOneTravelPatternsMax),
      coreConflict: zod.string(),
      whatYoureMissing: zod.string(),
      whyThisFits: zod.string(),
      tradeoffs: zod.string(),
      avoidIf: zod.array(zod.string()),
      nextMove: zod.string(),
      anchors: zod
        .array(zod.string())
        .min(getDecisionResponseResultJsonOneAnchorsMin)
        .max(getDecisionResponseResultJsonOneAnchorsMax),
      timingConfidence: zod.string(),
      stopDoingThis: zod.string(),
      usedSaveIds: zod.array(zod.number()),
    })
    .nullish(),
  savesSnapshot: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a decision
 */
export const DeleteDecisionParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteDecisionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Scrape metadata from a URL
 */
export const ScrapeUrlBody = zod.object({
  url: zod.string(),
});

export const ScrapeUrlResponse = zod.object({
  url: zod.string(),
  title: zod.string().nullish(),
  image: zod.string().nullish(),
  siteName: zod.string().nullish(),
  extractedText: zod.string().nullish(),
});